+++ lyc_flutter_project/lib/join/Provider/login_provider.dart
@@ -6,6 +6,7 @@ import 'package:lyc_flutter_project/common/const/data.dart';
 import 'package:lyc_flutter_project/common/dio/dio.dart';
 import 'package:lyc_flutter_project/mypage/model/profile.dart';
 import 'package:lyc_flutter_project/mypage/repository/mypage_repository.dart';
+import 'package:lyc_flutter_project/config/secret.dart';
 
 class LoginProvider extends ChangeNotifier {
   final DioProvider dioProvider;
@@ -20,7 +21,7 @@ class LoginProvider extends ChangeNotifier {
   Profile? _profile;
   bool _hasProfile = false;
 
-  static const String memberIdKey = 'jaskfjajesfn';
+
 
   get profile => _profile;
