+++ lyc_flutter_project/lib/join/screens/login_screen.dart
@@ -4,7 +4,7 @@ import 'package:lyc_flutter_project/data/app_color.dart';
 import 'package:lyc_flutter_project/screens/find_password_screen_1.dart';
 import 'package:lyc_flutter_project/Join/Screens/join_membership_screen_1.dart';
 import 'package:provider/provider.dart';
-import '../../findID/Screens/find_id_screen_1.dart';
+import '../../find_id/Screens/find_id_screen_1.dart';
 import '../../styles/app_text_style.dart';
 import '../../widget/Controller.dart';
 import '../Provider/login_provider.dart';
@@ -187,7 +187,7 @@ class LoginScreen extends StatelessWidget {
               alignment: Alignment.center,
               child: Center(
                 child: Padding(
-                  padding: EdgeInsets.fromLTRB(18, 12, 18, 12),
+                  padding: EdgeInsets.fromLTRB(16, 12, 16, 12),
                   child: TextField(
                     controller: controller.controller,
                     decoration: InputDecoration(
